# swagger.json
{
...
  "info": {
    "description": "Comprehensive API for bus payment system with offers, rewards, bookings, and refunds",
    "title": "Bus Payments & Rewards API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tokens/card/{card_id}/tokens": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of card tokens"
          },
          "404": {
            "description": "Credit card not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get card tokens",
        "tags": [
          "Card Tokens"
        ]
      }
    },
    "/api/tokens/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "card_id": {
                    "type": "integer"
                  },
                  "expires_in_hours": {
                    "default": 24,
                    "maximum": 8760,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "token_type": {
                    "default": "SINGLE_USE",
                    "enum": [
                      "SINGLE_USE",
                      "MULTI_USE",
                      "RECURRING"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "card_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Token created successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "Credit card not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create card token",
        "tags": [
          "Card Tokens"
        ]
      }
    },
    "/api/tokens/customer/{customer_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customer tokens"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get customer tokens",
        "tags": [
          "Card Tokens"
        ]
      }
    },
    "/api/tokens/{token_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token details"
          },
          "404": {
            "description": "Token not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get token details",
        "tags": [
          "Card Tokens"
        ]
      }
    },
    "/api/tokens/{token_id}/deactivate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token deactivated successfully"
          },
          "404": {
            "description": "Token not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Deactivate token",
        "tags": [
          "Card Tokens"
        ]
      }
    },
    "/api/tokens/{token_id}/validate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid"
          },
          "400": {
            "description": "Token is invalid or expired"
          },
          "404": {
            "description": "Token not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Validate token",
        "tags": [
          "Card Tokens"
        ]
      }
    },
...
    "/simulator/status": {
      "get": {
        "responses": {
          "200": {
            "description": "Simulator status information"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get simulator status",
        "tags": [
          "Integration"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://192.168.86.189:5001"
    }
  ]
}